Add copy value button to UI
