+++ indyscan-api-client/package.json
@@ -1,6 +1,6 @@
 {
   "name": "indyscan-api-client",
-  "version": "4.3.1",
+  "version": "4.4.0",
   "author": "Patrik Staš",
   "license": "ISC",
   "description": "IndyScan HTTP API client.",
