+++ indyscan-daemon-api-client/package.json
@@ -1,6 +1,6 @@
 {
   "name": "indyscan-daemon-api-client",
-  "version": "4.3.1",
+  "version": "4.4.0",
   "author": "Patrik Staš",
   "license": "ISC",
   "description": "IndyScan Daemon HTTP API client.",
