+++ indyscan-daemon-ui/package.json
@@ -1,6 +1,6 @@
 {
   "name": "indyscan-daemon-ui",
-  "version": "4.3.1",
+  "version": "4.4.0",
   "author": "Patrik Staš",
   "license": "ISC",
   "description": "UI to view and manage the state of indyscan-daemon.",
