+++ indyscan-txtype/package.json
@@ -1,6 +1,6 @@
 {
   "name": "indyscan-txtype",
-  "version": "4.3.1",
+  "version": "4.4.0",
   "description": "",
   "main": "index.js",
   "scripts": {
