+++ indyscan-webapp/components/BadgedValueDisplay/BadgedValueDisplay.scss
@@ -1,0 +1,27 @@
+  /* Tooltip container */
+.tooltip {
+  position: relative;
+  display: inline-block;
+}
+
+/* Tooltip text */
+.tooltip .tooltiptext {
+  visibility: visible;
+  width: 60px;
+  background-color: black;
+  color: #fff;
+  text-align: center;
+  padding: 5px 0;
+  border-radius: 6px;
+
+  /* Position the tooltip text - see examples below! */
+  position: absolute;
+  left: -50px;
+  bottom: 0.3em;
+  z-index: 1;
+}
+
+///* Show the tooltip text when you mouse over the tooltip container */
+//.tooltip:hover .tooltiptext {
+//  visibility: visible;
+//}
