+++ indyscan-webapp/package.json
@@ -1,6 +1,6 @@
 {
   "name": "indyscan-webapp",
-  "version": "4.3.1",
+  "version": "4.4.0",
   "description": "Web application to browse Hyperledger Indy blockchain transactions.",
   "main": "index.js",
   "scripts": {
