+++ version.json
@@ -1,5 +1,5 @@
 {
   "major": 4,
-  "minor": 3,
-  "patch": 1
+  "minor": 4,
+  "patch": 0
 }
